refactor(gridsat): rename netcdf_time module to time
