+++ src/goesdl/gridsat/utility.py
@@ -3,7 +3,7 @@ from netCDF4 import Dataset
 from ..geodesy import RectangularRegion
 from ..gridsat.netcdf_geodetic import GSLatLonGrid
 from ..gridsat.netcdf_image import GSImage
-from ..gridsat.netcdf_time import GSCoverageTime
+from .time import GSCoverageTime
 from .dataset_info import GSDatasetInfo
